event value convert

+++ Sources/Hackle/Core/Event/UserEventDispatcher.swift
@@ -164,7 +164,7 @@ extension UserEvents.Track {
         dto["eventTypeId"] = eventType.id
         dto["eventTypeKey"] = eventType.key
         if let value = event.value {
-            dto["value"] = value
+            dto["value"] = value.doubleValue
         }
         if let properties = event.properties {
             dto["properties"] = properties
